pyproject.toml
[project]
name = "tiny-dllm"
version = "0.1.0"
description = "Diffusion-style training for masked language models using ModernBERT"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers[torch]>=4.30.0",
    "datasets>=2.0.0",
    "sws-config",
]

[tool.uv]
dev-dependencies = []
